refact: error handling in philo_routine

+++ philo/inc/philo.h
@@ -88,7 +88,7 @@ void		mgo_putstr_fd(char *s, int fd);
 int			mgo_atoi(const char *str);
 int			error_with_msg(char *msg);
 void		*mgo_calloc(size_t count, size_t size);
-long int	set_time_ms(long int *dest);
+int			set_time_ms(long int *dest);
 int			sleep_shortly_ms(long int ms_sleep);
 int			print_philo_status(t_philo *philo, char *status);
 int			print_philo_died(t_philo *philo, long int ms_now);
